Print response time, waiting time, turnaround time and throughput. Fixed the limit on char array so that the last process can continue running until it's done according to the HW.

+++ process_producer.c
@@ -3,6 +3,7 @@
 #include "process_producer.h"
 #include "Sorter.h"
 
+
 void PrintProcessList(struct process process_list[]){
 	for(int x = 0; x < NUM_PROCESS; x++){
 		printf("%c: Service Time: %.1f Priority: %d Arrival Time: %d\n", process_list[x].name, process_list[x].service_time, process_list[x].priority, process_list[x].arrival_time);
@@ -11,7 +12,7 @@ void PrintProcessList(struct process process_list[]){
 
 void CreateProcesses(struct process** process_list){
 	char alphabet[26] = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-	int arrival_time; // [0 - 99]
+	float arrival_time; // [0 - 99]
 	float service_time; // [0.1 - 10]
 	int priority; // [1 - 4]
 	*process_list = calloc(NUM_PROCESS, sizeof(struct process));
